Bug 501616 - ArrayIndexOutOfBoundsException in CPPField.getFieldPosition
Use 16 bits to store field position.

+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/core/dom/ast/cpp/ICPPField.java
@@ -7,6 +7,7 @@
  *
  * Contributors:
  *     Doug Schaefer (IBM) - Initial API and implementation
+ *     Sergey Prigogin (Google)
  *******************************************************************************/
 package org.eclipse.cdt.core.dom.ast.cpp;
 
@@ -20,8 +21,10 @@ public interface ICPPField extends IField, ICPPMember, ICPPVariable {
     public static final ICPPField[] EMPTY_CPPFIELD_ARRAY = {};
     
     /**
-     * Returns the position of this field within its class owner's declared fields.
-	 * @since 6.0
+     * Returns the position of this field within its class owner's declared fields, or -1 if the position
+     * cannot be determined.
+     *
+	 * @since 6.1
 	 */
-    byte getFieldPosition();
+    int getFieldPosition();
 }
